refactor: 상수로 관리한다.
