+++ src/utils/index.ts
@@ -1,50 +1,10 @@
-import { TBeverageInfo } from 'type';
-
-export const BEVERAGE: TBeverageInfo[] = [
-  {
-    id: 0,
-    name: '에스프레소',
-    price: 600,
-    ingredients: [
-      { name: '컵', time: 1000 },
-      { name: '에스프레소', time: 2000 },
-    ],
-  },
-  {
-    id: 1,
-    name: '아메리카노',
-    price: 1000,
-    ingredients: [
-      { name: '컵', time: 1000 },
-      { name: '에스프레소', time: 2000 },
-      { name: '뜨거운 물', time: 1000 },
-    ],
-  },
-  {
-    id: 3,
-    name: '카페라떼',
-    price: 2000,
-    ingredients: [
-      { name: '컵', time: 1000 },
-      { name: '에스프레소', time: 2000 },
-      { name: '우유', time: 1000 },
-    ],
-  },
-  {
-    id: 4,
-    name: '우유',
-    price: 1000,
-    ingredients: [
-      { name: '컵', time: 1000 },
-      { name: '우유', time: 1000 },
-    ],
-  },
-];
+import { NUMBER_VALUE, MESSAGE } from 'constant';
 
 export const validateMoney = (money: number) => {
-  if (money <= 0) throw new Error('돈을 제대로 넣어주세요!');
-  if (money % 10)
-    throw new Error('10원으로 나누어떨어지는 금액만 입력 가능합니다!');
+  if (money <= NUMBER_VALUE.MINIMUM_MONEY)
+    throw new Error(MESSAGE.MINIMUM_MONEY);
+  if (money % NUMBER_VALUE.DIVISIBLE_NUMBER)
+    throw new Error(MESSAGE.DIVISIBLE_NUMBER);
 };
 
 export const clearAllTimers = (timers: number[]) => {
